refactor(project-pattern): updating constants
