+++ src/components/events/EventList.tsx
@@ -1,9 +1,9 @@
-import { EventListProps } from '@/types/components/EventsTypes'
+import { EventListProps } from '@/types/components/eventsTypes'
 import { EventCard } from './EventCard'
 
 export function EventList({ events }: EventListProps) {
   return (
-    <div className="flex flex-wrap gap-[5px]">
+    <div className='flex flex-wrap gap-[5px]'>
       {events.map((event, index) => (
         <EventCard event={event} key={index} />
       ))}
