--- src/components/hero/style.module.css
@@ -1,30 +1,0 @@
-@tailwind utilities;
-
-.text {
-  @apply opacity-0 text-center relative bottom-4 font-light animate-[text_3s_ease-in-out_0.5s_forwards];
-}
-
-@keyframes text {
-  10% {
-    transform: translateY(100px);
-  }
-  50% {
-    opacity: 0;
-  }
-  70% {
-    transform: translateY(0px);
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.line {
-  @apply scale-x-0 origin-right animate-[line-animation_0.5s_ease-in-out_1s_forwards]
-}
-
-@keyframes line-animation {
-  to {
-    transform: scaleX(1);
-  }
-}
