--- src/components/image_hero/index.ts
@@ -1,3 +1,0 @@
-import ImageHero from './ImageHero'
-
-export default ImageHero
