+++ src/components/section/Info.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { TextProps } from '@/types/components/SectionTypes'
+import { TextProps } from '@/types/components/sectionTypes'
 export function Info({ children }: TextProps) {
   return <div className='mt-2 font-poppins'>{children}</div>
 }
