+++ src/pages/_app.tsx
@@ -1,13 +1,14 @@
 import { appWithTranslation } from 'next-i18next'
+import { ThemeProvider } from '@/contexts/ThemeContext'
 import { AppProps } from 'next/app'
 import { Poppins } from 'next/font/google'
+import './globals.css'
+
 const poppins = Poppins({
   subsets: ['latin'],
   weight: ['100', '300', '400', '700', '600'],
   variable: '--font-poppins'
 })
-import './globals.css'
-import { ThemeProvider } from '@/contexts/ThemeContext'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
