+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { Header } from '@/components/header'
 import { Hero } from '@/components/hero'
-import { Newsletter } from '@/components/Newsletter'
+import { Newsletter } from '@/components/newsletter'
 import { SectionPrincipal } from '@/components/section/Section'
 import { NextPageContext } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
